+++ R/class-fun_cpt.R
@@ -73,3 +73,18 @@ whomademe <- function(x, ...) {
     parse(text = _) |>
     eval()
 }
+
+#' @rdname ls_pkgs
+#' @export
+#' @examples
+#' # List all model-fitting functions
+#' ls_models()
+#' 
+
+ls_models <- function() {
+  rlang::ns_env("tidychangepoint") |>
+    rlang::env_names() |>
+    stringr::str_subset("fit_") |>
+    stringr::str_subset("shift2$|region$|meanshift$", negate = TRUE) |>
+    sort()
+}